requirements.txt
torch>=2.3
torchvision>=0.18
numpy>=1.26
pydantic-settings>=2.4
typer[all]>=0.12
rich>=13.7
pyyaml>=6.0
onnx>=1.16
onnxruntime-gpu>=1.19; platform_system != "Darwin"
onnxruntime>=1.19; platform_system == "Darwin"
mlflow>=2.15
pytest>=8.0

